services/service_speak/requirements.txt
flask==2.3.3
kafka-python==2.0.2
coqui-tts
gtts
schedule==1.2.0
pymysql==1.1.0
